+++ src/components/icons/plus.tsx
@@ -1,0 +1,27 @@
+import type { IconProps } from "./types";
+
+export const IconPlus = (props: IconProps) => {
+  const {
+    className,
+    style,
+  } = props;
+
+  return (
+    <svg
+      width="13"
+      height="13"
+      viewBox="0 0 13 13"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+      style={style}
+    >
+      <path
+        d="M6.26758 0C6.81986 2.41411e-08 7.26758 0.447715 7.26758 1V5.25586H11.5234C12.0756 5.25596 12.5234 5.70364 12.5234 6.25586C12.5233 6.808 12.0756 7.25576 11.5234 7.25586H7.26758V11.5234C7.26756 12.0757 6.81985 12.5234 6.26758 12.5234C5.71539 12.5233 5.2676 12.0756 5.26758 11.5234V7.25586H1C0.447777 7.25586 0.000100509 6.80806 0 6.25586C0 5.70357 0.447715 5.25586 1 5.25586H5.26758V1C5.26758 0.447777 5.71538 0.000100493 6.26758 0Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+};
+
+export default IconPlus;